requirements.txt
# packages for API access
google-genai
anthropic[vertex]
google-cloud-translate==2.0.1

# general usage
python-dotenv